Code Refactor
