--- marketplace/ui/src/components/Membership/ServiceUsage.js
@@ -1,793 +1,0 @@
-import React, { useEffect, useState } from "react";
-import {
-  Tabs,
-  Table,
-  Input,
-  Select,
-  Button,
-  DatePicker,
-  Space,
-  InputNumber,
-  Row,
-  Col,
-  Typography,
-  Spin,
-} from "antd";
-import {
-  EditOutlined,
-  CheckOutlined,
-  CloseOutlined,
-  PlusOutlined,
-  DeleteOutlined,
-  LockOutlined,
-  CaretDownOutlined,
-} from "@ant-design/icons";
-import "./serviceUsage.css";
-import { actions as serviceUsageActions } from "../../contexts/serviceUsage/actions";
-import { actions as servicesActions } from "../../contexts/service/actions";
-import { actions as userAuthActions } from "../../contexts/authentication/actions";
-import { actions as membershipActions } from "../../contexts/membership/actions";
-import {
-  useServiceUsageDispatch,
-  useServiceUsageState,
-} from "../../contexts/serviceUsage";
-import {
-  useAuthenticateDispatch,
-  useAuthenticateState,
-} from "../../contexts/authentication";
-import {
-  useMembershipDispatch,
-  useMembershipState,
-} from "../../contexts/membership";
-import { useServiceDispatch, useServiceState } from "../../contexts/service";
-import moment from "moment";
-
-const statusOptions = [
-  { value: 1, label: "Requested" },
-  { value: 2, label: "Cancelled" },
-  { value: 3, label: "Completed" },
-];
-
-const limit = 10;
-const offset = 0;
-const query = "";
-
-const ServiceTable = () => {
-  const serviceUsageDispatch = useServiceUsageDispatch();
-  const serviceDispatch = useServiceDispatch();
-  const authUserDispatch = useAuthenticateDispatch();
-  const membershipDispatch = useMembershipDispatch();
-
-  // all api call states
-  const userCert = useAuthenticateState();
-  // const userListState = useUsersState();
-  const servicesState = useServiceState();
-  const membership = useMembershipState();
-  const serviceUsageState = useServiceUsageState();
-
-  const { isUsersLoading } = userCert;
-  const { isServicesLoading } = servicesState;
-  const { isPurchasedMembershipLoading } = membership;
-  const {
-    isServicesUsageLoading,
-    isCreateServiceUsageSubmitting,
-    isUpdateServicesUsageLoading,
-  } = serviceUsageState;
-  const IsLoading =
-    // isServicesLoading ||
-    isPurchasedMembershipLoading ||
-    isServicesUsageLoading ||
-    isUsersLoading ||
-    isCreateServiceUsageSubmitting ||
-    isUpdateServicesUsageLoading;
-
-  function transformData(data) {
-    const uniqueProducts = {};
-    const resultArray = [];
-
-    data.forEach((item) => {
-      const productId = item.productId;
-      const manufacturer = item.manufacturer;
-
-      // Check if the product ID is not already in the uniqueProducts object
-      if (!uniqueProducts[manufacturer]) {
-        uniqueProducts[manufacturer] = true; // Mark this product ID as seen
-        resultArray.push({ value: productId, label: manufacturer });
-      }
-    });
-
-    return resultArray;
-  }
-
-  const serviceUsageData = serviceUsageState?.servicesUsage;
-
-  const providerData = membership?.purchasedMemberships.map((item, index) => {
-    return { value: item.manufacturer, label: item.manufacturer };
-  });
-  const providerFilter = membership?.purchasedMemberships.map((item, index) => {
-    return { value: item.productId, label: item.manufacturer };
-  });
-  const defaultMembership = membership?.purchasedMemberships.map(
-    (item, index) => {
-      return { value: item.itemAddress, label: item.itemNumber };
-    }
-  );
-  const serviceListData = servicesState?.services?.map((item, index) => {
-    return { value: item.address, label: item.name };
-  });
-  const userListData = userCert?.users?.map((item, index) => {
-    return { value: item.userAddress, label: item.commonName };
-  });
-
-  const [providerList, setProviderList] = useState(providerData);
-  const [providerFilterList, setProviderFilterList] = useState(providerFilter);
-  const [membershipList, setMembershipList] = useState(defaultMembership);
-  const [serviceList, setServiceList] = useState(serviceListData);
-  const [userList, setUserList] = useState(userListData);
-  const [providerState, setProviderState] = useState("");
-  const [isAddRow, setIsAddRow] = useState(false)
-
-  useEffect(() => {
-    setServiceList(serviceListData);
-  }, [servicesState]);
-
-  useEffect(() => {
-    setProviderList(providerData);
-    setMembershipList(defaultMembership);
-    // setProviderFilterList(providerFilter)
-  }, [membership]);
-
-  useEffect(() => {
-    setTableData(serviceUsageData);
-  }, [serviceUsageState]);
-
-  useEffect(() => {
-    if (isAddRow) {
-      setTimeout(() => {
-        addRow()
-      }, 5000)
-    }
-  }, [isServicesUsageLoading])
-
-  useEffect(() => {
-    setUserList(userListData);
-  }, [userCert]);
-
-  const [isEdit, setIsEdit] = useState(false);
-  const [validationError, setValidationError] = useState(false);
-  const [activeTab, setActiveTab] = useState("booked");
-  const [tableData, setTableData] = useState([]);
-  const [page, setPage] = useState(1);
-  const [filterQuery, setFilterQuery] = useState({});
-  const username = userCert?.user?.commonName;
-  const organization = userCert?.user?.organization;
-  const userAddress = userCert?.user?.userAddress;
-
-  const newRowSchema = {
-    summary: "", //summary
-    serviceDate: "", //Date
-    providerComment: "", //comment
-    status: 1,
-    pricePaid: "", //price paid
-    editable: true,
-    itemId: "", //provider
-    serviceId: "", //service
-    paymentStatus: 1,
-
-    providerLastUpdated: userCert?.user?.userAddress, //user-address
-    providerLastUpdatedDate: new Date().getTime().toString(),
-  };
-
-  const queryOwner = `&owner=${userAddress}`;
-  useEffect(() => {
-    if (userAddress) {
-      if (activeTab === 'booked') {
-        serviceUsageActions.fetchBookedServicesUsage(
-          serviceUsageDispatch,
-          limit,
-          offset,
-          queryOwner
-        );
-      } else {
-        serviceUsageActions.fetchProvidedServicesUsage(
-          serviceUsageDispatch,
-          limit,
-          offset,
-          queryOwner
-        );
-      }
-    }
-    servicesActions.fetchService(serviceDispatch, 10, offset, query);
-    membershipActions.fetchPurchasedMemberships(membershipDispatch);
-    userAuthActions.fetchUsers(authUserDispatch);
-  }, [activeTab, userAddress]);
-
-  const handleTabChange = (key) => {
-    setFilterQuery({});
-    setPage(0);
-    setActiveTab(key);
-  };
-
-  const handleEditCancel = (key, bool, type, record) => {
-    // handling 3 functionality (i.e, edit, update & cancel) using its type
-
-    const UpdatePayloadKeys = [
-      "serviceDate",
-      "summary",
-      "status",
-      "paymentStatus",
-      "providerLastUpdated",
-      "providerComment",
-      "providerLastUpdatedDate",
-      "pricePaid",
-    ];
-    let updatedDataObj = {};
-    UpdatePayloadKeys.forEach((item, index) => {
-      if (
-        ["serviceDate", "providerLastUpdatedDate", "pricePaid"].includes(item)
-      ) {
-        updatedDataObj[item] = record[item].toString();
-      } else {
-        updatedDataObj[item] = record[item];
-      }
-    });
-
-    let updatedPayload = {};
-    updatedPayload["address"] = record.address;
-    updatedPayload["updates"] = updatedDataObj;
-
-    setIsEdit(bool);
-    const data = tableData.filter((item, index) => {
-      if (index === key) {
-        item["editable"] = bool;
-        return item;
-      } else if (type === "edit") {
-        item["editable"] = false;
-        return item;
-      }
-      return item;
-    });
-    setTableData(data);
-    if (type === "update") {
-      if (isEdit) {
-        // we have to use update api here
-        // uncomment api call for updating service usage
-        serviceUsageActions.UpdateServiceUsage(
-          serviceUsageDispatch,
-          updatedPayload
-        );
-      } else {
-        // we have to use create api here
-        updatedDataObj["itemId"] = record["itemId"];
-        updatedDataObj["serviceId"] = record["serviceId"];
-        serviceUsageActions.createServiceUsage(
-          serviceUsageDispatch,
-          updatedDataObj
-        );
-      }
-    }
-  };
-
-  const handleInputChange = (value, field, key) => {
-    if (field === "provider") {
-      let membershipData = membership?.purchasedMemberships
-        .filter(({ manufacturer }) => {
-          return manufacturer === value;
-        })
-        .map(({ itemAddress, itemNumber, manufacturer }) => {
-          return {
-            value: itemAddress,
-            label: itemNumber,
-            organization: manufacturer,
-          };
-        });
-      setMembershipList(membershipData);
-      // let serviceQuery = `&ownerOrganization=${value}`
-      servicesActions.fetchService(serviceDispatch, 10, offset, value);
-    } else if (field === "itemId") {
-      let data = tableData.filter((item, index) => {
-        if (index === key) {
-          item[field] = value ? value : "";
-          return item;
-        }
-        return item;
-      });
-      setTableData(data);
-    }
-    let data = tableData.filter((item, index) => {
-      if (index === key) {
-        item[field] = value ? value : "";
-        return item;
-      }
-      return item;
-    });
-    setTableData(data);
-  };
-
-  const addRow = () => {
-    setIsEdit(false);
-    let tableCopy = tableData.map((item, index) => {
-      item["editable"] = false;
-      return item;
-    });
-    let data = { ...newRowSchema };
-    data["key"] = tableCopy.length + 1;
-    setTableData([data, ...tableCopy]);
-    setIsAddRow(false);
-  }
-
-  const handleAddRow = () => {
-    // just add a new empty row in the tableData
-    if (page == 1) {
-      addRow();
-    } else {
-      setPage(1)
-      setIsAddRow(true)
-      serviceUsageActions.fetchBookedServicesUsage(serviceUsageDispatch, limit, offset, queryOwner);
-    }
-
-  };
-
-  const handleSave = () => {
-    console.log("tableData", tableData)
-    const data = tableData.map((item, index) => {
-      delete item["editable"];
-      delete item["key"];
-      delete item["provider"];
-      return item;
-    });
-    if (isEdit) {
-      serviceUsageActions.UpdateServiceUsage(serviceUsageDispatch, data);
-    } else {
-      serviceUsageActions.createServiceUsage(serviceUsageDispatch, data.at(0));
-    }
-  };
-
-  const handleDelete = (key) => {
-    let data = tableData.filter((item, index) => index !== key);
-    setTableData(data);
-  };
-
-  const handleValidation = (data) => {
-    const requiredFields = [
-      "summary",
-      "serviceDate",
-      "providerComment",
-      "status",
-      "pricePaid",
-      "serviceId",
-      "paymentStatus",
-      "providerLastUpdated",
-      "providerLastUpdatedDate",
-    ];
-
-    if (activeTab === "booked" || activeTab === "provided") {
-      if (requiredFields.every((field) => data[field] !== "" || null)) {
-        setValidationError(false);
-        return true;
-      }
-    }
-    setValidationError(true);
-    return false;
-  };
-
-  const handleQuery = (data, page) => {
-    let query1 = `&owner=${userAddress}`;
-    if (data.status) query1 += `&status=${data.status}`;
-
-    if (data?.Provider) {
-      const itemIds = membership?.purchasedMemberships
-        .filter((item) => item.manufacturer === data.Provider)
-        .map((item) => item.itemAddress);
-      if (itemIds.length > 0) {
-        query1 += `&itemId[]=${itemIds}`;
-      }
-    } else if (data?.User) {
-      query1 += `&providerLastUpdated=${data?.User}`
-    }
-    if (activeTab === 'booked') {
-      serviceUsageActions.fetchBookedServicesUsage(serviceUsageDispatch, limit, (page - 1) * limit, query1);
-    } else {
-      serviceUsageActions.fetchProvidedServicesUsage(serviceUsageDispatch, limit, (page - 1) * limit, query1);
-    }
-  }
-
-  const handleFilter = (value, key) => {
-    const data = { ...filterQuery, [key]: value };
-    setFilterQuery(data);
-    handleQuery(data, page)
-    setPage(1);
-  };
-
-  const columns = [
-    {
-      title: "User",
-      dataIndex: "providerLastUpdated",
-      key: "providerLastUpdated",
-      render: (text, record, index) => (
-        <span>
-          {record.editable && !isEdit ? (
-            <Select
-              placeholder="User"
-              defaultValue={username}
-              suffixIcon={
-                activeTab === "booked" ? (
-                  <LockOutlined />
-                ) : (
-                  <CaretDownOutlined />
-                )
-              }
-              disabled={activeTab === "booked"}
-              style={{ width: 120 }}
-              onChange={(value) =>
-                handleInputChange(value, "providerLastUpdated", index)
-              }
-              options={userList}
-            />
-          ) : (
-            <Typography style={{ color: "#061A6C" }}>
-              {userList.reduce((label, item) => {
-                if (item.value === text) {
-                  return item.label;
-                }
-                return label;
-              }, null)}
-            </Typography>
-          )}
-        </span>
-      ),
-    },
-    {
-      title: "Provider",
-      dataIndex: "provider",
-      key: "provider",
-      render: (text, record, index) => (
-        <span>
-          {record.editable && !isEdit ? (
-            <Select
-              placeholder="Provider"
-              defaultValue={activeTab === "provided" && organization}
-              suffixIcon={
-                activeTab === "provided" ? (
-                  <LockOutlined />
-                ) : (
-                  <CaretDownOutlined />
-                )
-              }
-              disabled={activeTab === "provided"}
-              style={{ width: 120 }}
-              onChange={(value, obj) => {
-                setProviderState(obj.value.toString());
-                handleInputChange(obj.label.toString(), "provider", index);
-              }}
-              options={transformData(membership?.purchasedMemberships)}
-            />
-          ) : (
-            <span>
-              {text}
-            </span>
-          )}
-        </span>
-      ),
-    },
-    {
-      title: "Membership ID",
-      dataIndex: "itemId",
-      key: "itemId",
-      render: (text, record, index) => (
-        <span>
-          {record.editable && !isEdit ? (
-            <Select
-              disabled={!record.provider}
-              placeholder="Membership ID"
-              suffixIcon={<CaretDownOutlined />}
-              style={{ width: 120 }}
-              onChange={(value, obj) => handleInputChange(obj.value, "itemId", index)}
-              options={membershipList}
-            />
-          ) : (
-            <Typography style={{ color: "#061A6C" }}>
-              {record.membershipNumber}
-            </Typography>
-          )}
-        </span>
-      ),
-    },
-    {
-      title: "Service",
-      dataIndex: "serviceId",
-      key: "serviceId",
-      render: (text, record, index) => (
-        <span>
-          {record.editable && !isEdit ? (
-            <Select
-              disabled={!!!providerState || isServicesLoading}
-              placeholder="Service"
-              suffixIcon={<CaretDownOutlined />}
-              style={{ width: 120 }}
-              onChange={(value) => handleInputChange(value, "serviceId", index)}
-              options={serviceList}
-            />
-          ) : (
-            <Typography style={{ color: "#061A6C" }}>
-              {record.serviceName}
-            </Typography>
-          )}
-        </span>
-      ),
-    },
-    {
-      title: "Summary",
-      dataIndex: "summary",
-      key: "summary",
-      render: (text, record, index) => (
-        <span>
-          {record.editable ? (
-            <Input
-              value={text}
-              suffix={<EditOutlined />}
-              placeholder="Summary"
-              onChange={(e) =>
-                handleInputChange(e.target.value, "summary", index)
-              }
-            />
-          ) : (
-            text
-          )}
-        </span>
-      ),
-    },
-    {
-      title: "Date",
-      dataIndex: "serviceDate",
-      key: "serviceDate",
-      render: (text, record, index) => (
-        <span>
-          {record.editable && !isEdit ? (
-            <DatePicker
-              // value={text ? moment(text, 'YYYY-MM-DD') : null}
-              onChange={(serviceDate, dateString) =>
-                handleInputChange(
-                  new Date(serviceDate).getTime().toString(),
-                  "serviceDate",
-                  index
-                )
-              }
-            />
-          ) : (
-            <Typography style={{ color: "#061A6C" }}>
-              {moment(text).format("MM-DD-YYYY")}
-            </Typography>
-          )}
-        </span>
-      ),
-    },
-    {
-      title: "Comments",
-      dataIndex: "providerComment",
-      key: "providerComment",
-      render: (text, record, index) => (
-        <span>
-          {record.editable ? (
-            <Input
-              value={text}
-              suffix={<EditOutlined />}
-              placeholder="Comments"
-              onChange={(e) =>
-                handleInputChange(e.target.value, "providerComment", index)
-              }
-            />
-          ) : (
-            text
-          )}
-        </span>
-      ),
-    },
-    {
-      title: "Status",
-      dataIndex: "status",
-      key: "status",
-      render: (text, record, index) => (
-        <span>
-          {record.editable ? (
-            <Select
-              value={text}
-              placeholder="Status"
-              suffixIcon={<CaretDownOutlined />}
-              // disabled={activeTab === "provided"}
-              style={{ minWidth: "100px" }}
-              onChange={(value) => handleInputChange(value, "status", index)}
-              options={statusOptions}
-            />
-          ) : (
-            <Typography style={{ color: "#061A6C" }}>
-              {statusOptions.reduce((label, item) => {
-                if (item.value === text) {
-                  return item.label;
-                }
-                return label;
-              }, null)}
-            </Typography>
-          )}
-        </span>
-      ),
-    },
-    {
-      title: "Price Paid",
-      dataIndex: "pricePaid",
-      key: "pricePaid",
-      render: (text, record, index) => (
-        <span>
-          {record.editable && !isEdit ? (
-            <InputNumber
-              keyboard={true}
-              className="w-36"
-              addonAfter={<EditOutlined />}
-              min={0}
-              type="number"
-              controls={false}
-              value={parseInt(text)}
-              placeholder="Price Paid"
-              onChange={(value) =>
-                handleInputChange(value && value.toString(), "pricePaid", index)
-              }
-            />
-          ) : (
-            <Typography style={{ color: "#061A6C" }}>{text}</Typography>
-          )}
-        </span>
-      ),
-    },
-    {
-      title: "",
-      dataIndex: "actions",
-      key: "actions",
-      render: (_, record, index) => (
-        <Space size="middle">
-          {record.editable ? (
-            <>
-              <Button
-                type="primary"
-                icon={<CheckOutlined />}
-                disabled={!handleValidation(record)}
-                onClick={() => handleEditCancel(index, false, "update", record)}
-              />
-              {isEdit && (
-                <Button
-                  type="default"
-                  icon={<CloseOutlined />}
-                  onClick={() =>
-                    handleEditCancel(index, false, "cancel", record)
-                  }
-                />
-              )}
-            </>
-          ) : (
-            <Button
-              type="primary"
-              icon={<EditOutlined />}
-              disabled={!handleValidation(record) || validationError}
-              onClick={() => handleEditCancel(index, true, "edit", record)}
-            />
-          )}
-          {record.editable && !isEdit && (
-            <Button
-              type="danger"
-              icon={<DeleteOutlined />}
-              onClick={() => handleDelete(index)}
-            />
-          )}
-        </Space>
-      ),
-    },
-  ];
-
-  const tabOptions = [
-    {
-      key: "booked",
-      label: "Booked",
-    },
-    {
-      key: "provided",
-      label: "Provided",
-    },
-  ];
-
-  const dataLen = serviceUsageState?.servicesUsage?.length;
-  const paginationConfig = {
-    current: page,
-    pageSize: 10, // Number of items to display per page
-    total: dataLen == 10 ? (page + 1) * limit + 1 : page * limit, // Total number of items
-    showSizeChanger: false, // Allow users to change the page size
-    position: ["bottomCenter"],
-  };
-
-  const handlePaginationChange = (CPage) => {
-    setPage(CPage.current);
-    let page = CPage.current;
-    if (activeTab === 'booked') {
-      handleQuery(filterQuery, page)
-    } else {
-      handleQuery(filterQuery, page)
-    }
-  };
-
-  const activeTabCheck = activeTab === "booked" ? "Provider" : "User";
-
-  return (
-    <>
-      <Row className="mt-2">
-        <Col span={22} className="m-auto">
-          <Tabs
-            activeKey={activeTab}
-            items={tabOptions}
-            onChange={handleTabChange}
-          />
-        </Col>
-        <Col
-          className="flex justify-between absolute right-20 mt-2 z-10"
-          span={4}
-        >
-          <Button
-            type="primary"
-            icon={<PlusOutlined />}
-            onClick={handleAddRow}
-            disabled={validationError || IsLoading || (tableData && !tableData[0]?.address)}
-          >
-            Add Service Use
-          </Button>
-        </Col>
-      </Row>
-      <Row>
-        <Col span={22} className="m-auto flex justify-between">
-          <Typography.Title level={4} style={{ color: "#061A6C" }}>
-            Service Usage
-          </Typography.Title>
-          <span className="service-filter">
-            <Select
-              placeholder={activeTabCheck}
-              suffixIcon={<CaretDownOutlined />}
-              style={{ width: 120 }}
-              disabled={IsLoading}
-              value={filterQuery[activeTabCheck]}
-              onChange={(value, obj) => {
-                handleFilter(activeTab == 'booked' ? obj.label : value, activeTabCheck);
-              }}
-              options={
-                activeTab === "booked"
-                  ? transformData(membership?.purchasedMemberships)
-                  : userList
-              }
-            />
-            <Select
-              placeholder="Status"
-              className="ml-2"
-              suffixIcon={<CaretDownOutlined />}
-              disabled={IsLoading}
-              style={{ width: 120 }}
-              value={filterQuery["status"]}
-              onChange={(value) => {
-                handleFilter(value, "status");
-              }}
-              options={statusOptions}
-            />
-          </span>
-        </Col>
-      </Row>
-      <Row>
-        <Col span={22} className="m-auto">
-          <Table
-            columns={columns}
-            dataSource={tableData}
-            pagination={dataLen <= 10 && page == 0 ? false : paginationConfig}
-            loading={IsLoading}
-            rowKey="key"
-            onChange={handlePaginationChange} // Add this line to handle pagination changes
-          />
-        </Col>
-      </Row>
-    </>
-  );
-};
-
-export default ServiceTable;
